langgraph-templates/sql-agent-workflow/pyproject.toml
[project]
name = "sql-agent-workflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipython>=8.31.0",
    "langchain-community>=0.3.14",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.62",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
